+++ libc/stdio/itoa.c
@@ -1,10 +1,11 @@
 #include "stdio.h"
+#include <stdint.h>
 
-char *itoa(int value, char *str, int base) {
+char *itoa(int64_t value, char *str, int base) {
   if (base == 10 && value < 0) {
     str[0] = '-';
-    utoa((unsigned int)(-value), str + 1, base);
+    utoa((uint64_t)(-value), str + 1, base);
     return str;
   }
-  return utoa((unsigned int)value, str, base);
+  return utoa((uint64_t)value, str, base);
 }
